Move curlutils to new http library.
