+++ http/curlutils.c
