+++ http/curlutils.h
